refactor: Client Socket 마이그레이션
- socket 연결을 메서드로 호출하는 방식으로 변경
- 체결내역 저장 기능 추가

+++ src/main/java/com/kr/economy/tradebatch/trade/application/OrderInCashCommand.java
@@ -19,6 +19,6 @@ public class OrderInCashCommand {
     private OrderDvsnCode orderDvsnCode;
     private Float sharePrice;
     private Float orderPrice;
-    private Float qty;
+    private int qty;
     private KisOrderDvsnCode kisOrderDvsnCode;
 }
